requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
sentence-transformers==2.7.0
pdfminer.six==20231228
requests==2.31.0
torch==2.3.0
pydantic==2.7.1